+++ src/main/java/com/market/marketplace/dto/CartResponse.java
@@ -2,11 +2,12 @@ package com.market.marketplace.dto;
 
 import com.market.marketplace.model.Cart;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartResponse (
         List<CartItemResponse> items,
-        Double total
+        BigDecimal total
 ) {
     public static CartResponse from(Cart cart) {
         return new CartResponse(
